refactor: extract runtime output helpers
